Fix styles.
